+++ CHANGELOG.md
@@ -5,6 +5,8 @@
   note link dialog (for [#1828](https://github.com/pbek/QOwnNotes/issues/1828#issuecomment-663835422))
 - fixed warnings when encrypting notes if application is built with Qt 5.14 or newer
   (for [#1832](https://github.com/pbek/QOwnNotes/issues/1832))
+- changes to individual notes are now also ignored if the setting to ignore all
+  external note folder changes is active
 
 ## 20.7.6
 - the note link dialog can now be resized smaller than the height needed for its 2nd tab